refactor: update RouteTabItem to include enabled prop and remove legacy tab-config file
